+++ packages/php-laravel/src/SSOClientServiceProvider.php
@@ -9,8 +9,14 @@ class SSOClientServiceProvider extends ServiceProvider
 {
     public function boot()
     {
+        $this->mergeConfigFrom(__DIR__ . '/sso.php', 'sso');
+
+        if (config('sso.register_routes', true)) {
+            $this->loadRoutesFrom(__DIR__ . '/routes/sso.php');
+        }
+
         $this->publishes([
-            __DIR__ . '/SSOController.php' => App::path('Http/Controllers/SSO/SSOController.php'),
+            __DIR__ . '/../stubs/SSOController.php' => App::path('Http/Controllers/SSO/SSOController.php'),
             __DIR__ . '/sso.php' => config_path('sso.php'),
         ], 'sso-config');
     }